// packages/runtime-core/src/types/data-type.ts
enum BuiltInDataType {
  Boolean = 'boolean',
  Integer = 'integer',
  Float = 'float',
  String = 'string',
  Text = 'text',
  Enum = 'enum',
  MultiEnum = 'multi-enum',
  Date = 'date',
  OneToOne = 'o2o',
  OneToMany = 'o2m',
  ManyToMany = 'm2m',
  ManyToOne = 'm2o',
}

export { BuiltInDataType };